refactor: convert devicepreferences:create command to yargs
